# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface HealthStatus {
  status: string;
}

export type TradeResult = typeof TradeResult[keyof typeof TradeResult];


export const TradeResult = {
  WIN: 'WIN',
  LOSS: 'LOSS',
  BREAKEVEN: 'BREAKEVEN',
} as const;

export type TradeBeforeAnswers = { [key: string]: unknown };

export type TradeDuringAnswers = { [key: string]: unknown };

export type TradeAfterAnswers = { [key: string]: unknown };

export interface Trade {
  id: number;
  createdAt: string;
  result: TradeResult;
  /**
     * @minimum 0
     * @maximum 100
     */
  disciplineScore: number;
  beforeAnswers: TradeBeforeAnswers;
  duringAnswers: TradeDuringAnswers;
  afterAnswers: TradeAfterAnswers;
  screenshots: string[];
  /** @nullable */
  pnl: number | null;
  mistake: string;
}

export type TradeInputResult = typeof TradeInputResult[keyof typeof TradeInputResult];


export const TradeInputResult = {
  WIN: 'WIN',
  LOSS: 'LOSS',
  BREAKEVEN: 'BREAKEVEN',
} as const;

export type TradeInputBeforeAnswers = { [key: string]: unknown };

export type TradeInputDuringAnswers = { [key: string]: unknown };

export type TradeInputAfterAnswers = { [key: string]: unknown };

export interface TradeInput {
  result: TradeInputResult;
  /**
     * @minimum 0
     * @maximum 100
     */
  disciplineScore: number;
  beforeAnswers: TradeInputBeforeAnswers;
  duringAnswers: TradeInputDuringAnswers;
  afterAnswers: TradeInputAfterAnswers;
  screenshots?: string[];
  /** @nullable */
  pnl?: number | null;
  mistake?: string;
}

export type TradeUpdateResult = typeof TradeUpdateResult[keyof typeof TradeUpdateResult];


export const TradeUpdateResult = {
  WIN: 'WIN',
  LOSS: 'LOSS',
  BREAKEVEN: 'BREAKEVEN',
} as const;

export type TradeUpdateBeforeAnswers = { [key: string]: unknown };

export type TradeUpdateDuringAnswers = { [key: string]: unknown };

export type TradeUpdateAfterAnswers = { [key: string]: unknown };

export interface TradeUpdate {
  result?: TradeUpdateResult;
  /**
     * @minimum 0
     * @maximum 100
     */
  disciplineScore?: number;
  beforeAnswers?: TradeUpdateBeforeAnswers;
  duringAnswers?: TradeUpdateDuringAnswers;
  afterAnswers?: TradeUpdateAfterAnswers;
  screenshots?: string[];
  /** @nullable */
  pnl?: number | null;
  mistake?: string;
}

export type AnalyticsSummaryPnlSeriesItem = {
  label: string;
  value: number;
};

export type AnalyticsSummaryDisciplineSeriesItem = {
  label: string;
  value: number;
};

export interface AnalyticsSummary {
  todayTrades: number;
  todayPnl: number;
  winRate: number;
  disciplineScore: number;
  currentStreak: number;
  totalTrades: number;
  pnlSeries: AnalyticsSummaryPnlSeriesItem[];
  disciplineSeries: AnalyticsSummaryDisciplineSeriesItem[];
}

export interface RequestUploadUrlInput {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlMetadata = {
  name: string;
  size: number;
  contentType: string;
};

export interface RequestUploadUrl {
  uploadURL: string;
  objectPath: string;
  metadata: RequestUploadUrlMetadata;
}

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};