# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: trades
    description: Trading discipline records
  - name: analytics
    description: Trading performance summaries
  - name: storage
    description: Screenshot uploads
  - name: Auth
    description: Browser authentication endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trade records
      responses:
        "200":
          description: Trade records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Create a trade record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Created trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a trade record
      parameters:
        - $ref: "#/components/parameters/TradeId"
      responses:
        "200":
          description: Trade record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Trade not found
    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a trade record
      parameters:
        - $ref: "#/components/parameters/TradeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade record
      parameters:
        - $ref: "#/components/parameters/TradeId"
      responses:
        "204":
          description: Deleted
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get dashboard performance summary
      responses:
        "200":
          description: Performance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned screenshot upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlInput"
      responses:
        "200":
          description: Presigned upload details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrl"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Authentication state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start browser login
      parameters:
        - name: returnTo
          in: query
          required: false
          schema: { type: string }
      responses:
        "302":
          description: Redirect to the identity provider
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete browser login
      responses:
        "302":
          description: Redirect back to the application
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the browser session
      parameters:
        - name: returnTo
          in: query
          required: false
          schema: { type: string }
      responses:
        "302":
          description: Redirect to the identity provider logout
components:
  parameters:
    TradeId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id: { type: string }
        email: { type: ["string", "null"] }
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        profileImageUrl: { type: ["string", "null"] }
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Trade:
      type: object
      required: [id, createdAt, result, disciplineScore, beforeAnswers, duringAnswers, afterAnswers, screenshots, pnl, mistake]
      properties:
        id: { type: number }
        createdAt: { type: string, format: date-time }
        result: { type: string, enum: [WIN, LOSS, BREAKEVEN] }
        disciplineScore: { type: number, minimum: 0, maximum: 100 }
        beforeAnswers: { type: object, additionalProperties: true }
        duringAnswers: { type: object, additionalProperties: true }
        afterAnswers: { type: object, additionalProperties: true }
        screenshots:
          type: array
          items: { type: string }
        pnl: { type: ["number", "null"] }
        mistake: { type: string }
    TradeInput:
      type: object
      required: [result, disciplineScore, beforeAnswers, duringAnswers, afterAnswers]
      properties:
        result: { type: string, enum: [WIN, LOSS, BREAKEVEN] }
        disciplineScore: { type: number, minimum: 0, maximum: 100 }
        beforeAnswers: { type: object, additionalProperties: true }
        duringAnswers: { type: object, additionalProperties: true }
        afterAnswers: { type: object, additionalProperties: true }
        screenshots:
          type: array
          items: { type: string }
        pnl: { type: ["number", "null"] }
        mistake: { type: string }
    TradeUpdate:
      type: object
      properties:
        result: { type: string, enum: [WIN, LOSS, BREAKEVEN] }
        disciplineScore: { type: number, minimum: 0, maximum: 100 }
        beforeAnswers: { type: object, additionalProperties: true }
        duringAnswers: { type: object, additionalProperties: true }
        afterAnswers: { type: object, additionalProperties: true }
        screenshots:
          type: array
          items: { type: string }
        pnl: { type: ["number", "null"] }
        mistake: { type: string }
    AnalyticsSummary:
      type: object
      required: [todayTrades, todayPnl, winRate, disciplineScore, currentStreak, totalTrades, pnlSeries, disciplineSeries]
      properties:
        todayTrades: { type: number }
        todayPnl: { type: number }
        winRate: { type: number }
        disciplineScore: { type: number }
        currentStreak: { type: number }
        totalTrades: { type: number }
        pnlSeries:
          type: array
          items:
            type: object
            required: [label, value]
            properties:
              label: { type: string }
              value: { type: number }
        disciplineSeries:
          type: array
          items:
            type: object
            required: [label, value]
            properties:
              label: { type: string }
              value: { type: number }
    RequestUploadUrlInput:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string }
        size: { type: number }
        contentType: { type: string }
    RequestUploadUrl:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
        metadata:
          type: object
          required: [name, size, contentType]
          properties:
            name: { type: string }
            size: { type: number }
            contentType: { type: string }